add CAD models for boat
includes all CAD files for the entire project. All solidworks. Hull mold files are configurations inside the main pontoon geometry. Main assembly is titled "Main Assy". Consider Main Assy to be the master for true to life boat geometry as built. Some other parts in this upload may be unused, obsolete, or contain configurations that are not relevant anymore.